Wake up the consumer poll loop on certain signals.
On receiving SIGINT or SIGTERM wakeup the consumer to exit the poll
method straight away.

Prior to this the consumer would be allowed to finish its poll
however the process would get SIGKILLED regardless as instance got
terminated by EC2.

This leads to an unclean shutdown and consequently the consumer
does not get the chance to withdraw gracefully from the consumer
group. The result is threshing in the rebalancing world.